Move `music-constants.ts` to `src/constants/`
